+++ src/Controllers/Admin/UserVoucherController.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Engelsystem\Controllers\Admin;
 
-use Carbon\Carbon;
 use Engelsystem\Config\Config;
 use Engelsystem\Controllers\BaseController;
 use Engelsystem\Controllers\HasUserNotifications;
 use Engelsystem\Helpers\Authenticator;
+use Engelsystem\Helpers\UserVouchers;
 use Engelsystem\Http\Exceptions\HttpNotFound;
 use Engelsystem\Http\Redirector;
 use Engelsystem\Http\Request;
@@ -58,7 +58,7 @@ class UserVoucherController extends BaseController
                 'userdata' => $user,
                 'gotVouchers' => $user->state->got_voucher ?? 0,
                 'forceActive' => $user->state->force_active && config('enable_force_active'),
-                'eligibleVoucherCount' => $this->eligibleVoucherCount($user),
+                'eligibleVoucherCount' => UserVouchers::eligibleVoucherCount($user),
             ]
         );
     }
@@ -89,51 +89,5 @@ class UserVoucherController extends BaseController
 
         return $this->redirect->to('/users?action=view&user_id=' . $user->id);
         // TODO Once User_view.php gets removed, change this to withView + getNotifications
-    }
-
-    private function eligibleVoucherCount(User $user): int
-    {
-        $voucherSettings = config('voucher_settings');
-        $start = $voucherSettings['voucher_start']
-            ? Carbon::createFromFormat('Y-m-d', $voucherSettings['voucher_start'])->setTime(0, 0)
-            : null;
-
-        $shiftEntries = $user->shiftEntries()
-            ->join('shifts', 'shift_entries.shift_id', '=','shifts.id')
-            ->whereDate('shifts.end', '<', Carbon::now())
-            ->whereDate('shifts.start', '>=', $start ?: 0)
-            ->whereNull('freeloaded_by')
-            ->get();
-        $worklogs = $user->worklogs()
-            ->whereDate('worked_at', '>=', $start ?: 0)
-            ->whereDate('worked_at', '<=', Carbon::now())
-            ->get();
-        $shiftsCount =
-            $shiftEntries->count()
-            + $worklogs->count();
-
-        $shiftsTime = 0;
-        foreach ($shiftEntries as $shiftEntry) {
-            $shiftsTime += $shiftEntry->shift->start->diffInHours($shiftEntry->shift->end);
-        }
-        foreach ($worklogs as $worklog) {
-            $shiftsTime += $worklog->hours;
-        }
-
-        $vouchers = $voucherSettings['initial_vouchers'];
-        if ($voucherSettings['shifts_per_voucher']) {
-            $vouchers += $shiftsCount / $voucherSettings['shifts_per_voucher'];
-        }
-        if ($voucherSettings['hours_per_voucher']) {
-            $vouchers += $shiftsTime / $voucherSettings['hours_per_voucher'];
-        }
-
-        $vouchers -= $user->state->got_voucher;
-        $vouchers = floor($vouchers);
-        if ($vouchers <= 0) {
-            return 0;
-        }
-
-        return (int) $vouchers;
     }
 }
